requirements.txt
opencv-python~=4.9.0.80
tensorflow~=2.15.0
numpy~=1.26.3
flask~=3.0.1
werkzeug~=3.0.1
inflect~=7.0.0
pillow~=10.2.0